few minor changes in merged Technology+Associate+Question Module
